docker/app/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
geojson
requests
pyyaml
